# outputs/english-platform-v1-1/openapi-v1.yaml
openapi: 3.1.0
...
info:
  title: 个性化英语学习平台 API
  version: 1.1.0
  summary: General 与 TOEFL 核心学习闭环的官方 Web API 契约
  description: |
    本契约是 V1.1 服务端与官方 Web 端之间的唯一接口真相源。

    所有租户资源均位于 `/api/v1/tenants/{tenantId}/...`。服务端必须先验证当前
    用户的有效租户成员关系，再在数据库事务内设置租户、用户和成员上下文，由
    PostgreSQL RLS 执行第二道隔离。调用者无权知道目标资源是否存在时统一返回
    `404`；资源本身对调用者可见、但当前操作不被允许时才返回 `403`。

    日期时间字段必须是 UTC RFC 3339（以 `Z` 结尾），标识符均为 UUID。列表使用
    不透明游标分页。错误响应使用 RFC 7807 `application/problem+json`。

    官方 Web 端使用 HttpOnly Access Cookie。所有有状态写请求还必须携带
    `X-CSRF-Token`，其值与可读 CSRF Cookie 组成双提交校验。标记了
    `Idempotency-Key` 的命令在相同主体、租户、路由和请求体下重试时必须返回首次
    结果；同键不同请求体返回 `409`。自动保存使用 `If-Match` 与响应 `ETag`，并以
    `revision` 提供可观测的乐观并发版本。
  contact:
    name: English Platform Engineering
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Same-origin official Web application
...
security:
  - accessCookie: []
paths:
  /api/v1/auth/csrf:
    get:
      tags: [Authentication]
      summary: Bootstrap CSRF protection
      operationId: getCsrfToken
      security: []
      responses:
        '200':
          description: CSRF token returned and mirrored in a SameSite CSRF cookie
          headers:
            Set-Cookie:
              description: Non-HttpOnly CSRF cookie; never contains an authentication secret
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfTokenResponse'
              example:
                token: 2f75eac8-62e5-49fb-bcc3-4f260141c6aa
                expiresAt: '2026-07-10T13:00:00Z'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Sign in to the official Web application
      description: Origin and Fetch Metadata validation are required in addition to the CSRF token.
      operationId: login
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              identifier: learner001
              password: correct-horse-battery-staple
      responses:
        '200':
          description: Authenticated; rotating Access and Refresh cookies are set
          headers:
            Set-Cookie:
              $ref: '#/components/headers/AuthSetCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/password:
    post:
      tags: [Authentication]
      summary: Replace the current or temporary password
      operationId: changePassword
      security:
        - accessCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed and all other sessions revoked
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CsrfForbidden'
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate the session cookies
      description: Refresh-token reuse revokes the entire token family and returns 401.
      operationId: refreshSession
      security:
        - refreshCookie: []
          csrfHeader: []
      responses:
        '200':
          description: Session rotated
          headers:
            Set-Cookie:
              $ref: '#/components/headers/AuthSetCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CsrfForbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/session:
    delete:
      tags: [Authentication]
      summary: Sign out and revoke the active refresh-token family
      operationId: logout
      security:
        - accessCookie: []
          csrfHeader: []
      responses:
        '204':
          description: Session revoked; authentication cookies are expired
          headers:
            Set-Cookie:
              $ref: '#/components/headers/AuthSetCookie'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CsrfForbidden'
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: Short-lived, HttpOnly, Secure, SameSite=Lax signed access cookie.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotating, HttpOnly, Secure, SameSite=Strict opaque refresh cookie; only its hash is stored.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit token that must match the readable csrf_token cookie on every state-changing request.
...
  headers:
    AuthSetCookie:
      description: One or more Set-Cookie values for access_token and refresh_token rotation or expiry.
      schema:
        type: string
...
  responses:
    ValidationProblem:
      description: Request validation failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://api.example.cn/problems/validation-error
            title: 请求参数无效
            status: 400
            detail: 一个或多个字段未通过校验。
            instance: /api/v1/tenants/d61f2d08-a85e-4bbc-b860-dfdcbec68dc9/student/progress
            requestId: 0198a371-8261-7e53-9905-c81722fc08db
            errors:
              - field: to
                code: date_range_invalid
                message: to 必须不早于 from。
    Unauthorized:
      description: Authentication is absent, expired, revoked, or refresh-token reuse was detected.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    CsrfForbidden:
      description: CSRF, Origin, or Fetch Metadata validation failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          description: Seconds until another request should be attempted.
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    Uuid:
      type: string
      format: uuid
      description: Server-generated UUIDv7; clients treat it as an opaque UUID on the wire.
      example: 0198a352-01be-75a6-8e96-8ce78921799a
    UtcDateTime:
      type: string
      format: date-time
      pattern: '^(?:[0-9]{4})-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12][0-9]|3[01])T(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9](?:\.[0-9]+)?Z$'
      description: UTC RFC 3339 timestamp serialized from PostgreSQL timestamptz; offsets other than Z are rejected.
      example: '2026-07-10T12:00:00Z'
...
    ProblemFieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    ProblemDetails:
      type: object
      description: RFC 7807 problem detail with stable platform extensions.
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        requestId:
          $ref: '#/components/schemas/Uuid'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ProblemFieldError'
      additionalProperties: true
    CsrfTokenResponse:
      type: object
      additionalProperties: false
      required: [token, expiresAt]
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 256
        expiresAt:
          $ref: '#/components/schemas/UtcDateTime'
    LoginRequest:
      type: object
      additionalProperties: false
      required: [identifier, password]
      properties:
        identifier:
          type: string
          minLength: 3
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        newPassword:
          type: string
          format: password
          minLength: 12
          maxLength: 128
    PlatformRole:
      type: string
      enum: [none, super_admin]
    CurrentUser:
      type: object
      additionalProperties: false
      required: [id, email, loginName, displayName, platformRole, mustChangePassword, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        email:
          type: [string, 'null']
          format: email
        loginName:
          type: [string, 'null']
          maxLength: 64
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        platformRole:
          $ref: '#/components/schemas/PlatformRole'
        mustChangePassword:
          type: boolean
        createdAt:
          $ref: '#/components/schemas/UtcDateTime'
    SessionResponse:
      type: object
      additionalProperties: false
      required: [user, accessExpiresAt, refreshExpiresAt]
      properties:
        user:
          $ref: '#/components/schemas/CurrentUser'
        accessExpiresAt:
          $ref: '#/components/schemas/UtcDateTime'
        refreshExpiresAt:
          $ref: '#/components/schemas/UtcDateTime'
...